requirements.txt
fastapi
uvicorn
authlib
httpx
itsdangerous
python-dotenv